Integrated Stripe Payment

+++ BookshopClient/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -1,10 +1,12 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, ElementRef, Input, OnInit, ViewChild, } from '@angular/core';
 import { ShoppingCartService } from 'src/app/shopping-cart/shopping-cart.service';
 import { CheckoutService } from '../checkout.service';
 import { ToastrService } from 'ngx-toastr';
 import { NavigationExtras, Router } from '@angular/router';
 import { IShoppingCart } from 'src/app/shared/models/shoppingCart';
 import { FormGroup } from '@angular/forms';
+import { loadStripe } from '@stripe/stripe-js';
+import { firstValueFrom } from 'rxjs';
 
 @Component({
   selector: 'app-checkout-payment',
@@ -13,29 +15,79 @@ import { FormGroup } from '@angular/forms';
 })
 export class CheckoutPaymentComponent implements OnInit {
   @Input() checkoutForm!: FormGroup;
+  @ViewChild('cardNumber', {static: true}) cardNumberElement!: ElementRef;
+  @ViewChild('cardExpiry', {static: true}) cardExpiryElement!: ElementRef;
+  @ViewChild('cardCvc', {static: true}) cardCvcElement!: ElementRef;
+  stripe: any;
+  cardNumber: any;
+  cardExpiry: any;
+  cardCvc: any;
+  cardNumberComplete = false;
+  cardExpiryComplete = false;
+  cardCvcComplete = false;
+  cardErrors: any;
+  loading = false;
 
   constructor(private shoppingCartService: ShoppingCartService, private checkoutService: CheckoutService,
               private toastr: ToastrService, private router: Router) {}
 
   ngOnInit(): void {
-    
+    loadStripe('pk_test_51POGAMLj5bm6aVXKuCtCcsmnC865di023elKnLlTv7kkSXN0wgcLLaKuQ7NBeIcDIs4elNVIwlkDS5A9mLqCIZeT00adrQVcFu').then(stripe => {
+      this.stripe = stripe;
+      const elements = stripe?.elements();
+      if (elements) {
+        this.cardNumber = elements.create('cardNumber');
+        this.cardNumber.mount(this.cardNumberElement.nativeElement);
+        this.cardNumber.on('change', (event: any) => {
+          this.cardNumberComplete = event.complete;
+          if(event.error)
+            this.cardErrors = event.error.message;
+          else
+            this.cardErrors = null;
+        })
+
+        this.cardExpiry = elements.create('cardExpiry');
+        this.cardExpiry.mount(this.cardExpiryElement.nativeElement);
+        this.cardExpiry.on('change', (event: any) => {
+          this.cardExpiryComplete = event.complete;
+          if(event.error)
+            this.cardErrors = event.error.message;
+          else
+            this.cardErrors = null;
+        })
+
+        this.cardCvc = elements.create('cardCvc');
+        this.cardCvc.mount(this.cardCvcElement.nativeElement);
+        this.cardCvc.on('change', (event: any) => {
+          this.cardCvcComplete = event.complete;
+          if(event.error)
+            this.cardErrors = event.error.message;
+          else 
+            this.cardErrors = null;
+        })
+      }
+    })
   }
 
-  submitOrder() {
+  async submitOrder() {
+    this.loading = true;
     const shoppingCart = this.shoppingCartService.getCurrentShoppingCartValue();
-    const orderToCreate = this.getOrderToCreate(shoppingCart!);
-    this.checkoutService.createOrder(orderToCreate).subscribe({
-      next: order => {
-        this.toastr.success('Uspešno ste izvršili porudžbinu');
+    try {
+      const createdOrder = await this.createOrder(shoppingCart!);
+      const paymentResult = await this.confirmPaymentWithStripe(shoppingCart!);
+
+      if(paymentResult.paymentIntent) {
         this.shoppingCartService.deleteShoppingCart(shoppingCart!);
-        const navigationExtras: NavigationExtras = {state: order};
+        const navigationExtras: NavigationExtras = {state: createdOrder};
         this.router.navigate(['checkout/success'], navigationExtras);
-      },
-      error: error => {
-        this.toastr.error(error.message);
-        console.log(error);
+      } else {
+        this.toastr.error('Neuspešno plaćanje. ' + paymentResult.error.message);
       }
-    });
+      this.loading = false;
+    } catch (error) {
+      console.log(error);
+      this.loading = false;
+    }
   }
 
   private getOrderToCreate(shoppingCart: IShoppingCart) {
@@ -46,4 +98,19 @@ export class CheckoutPaymentComponent implements OnInit {
     }
   }
 
+  private async createOrder(shoppingCart: IShoppingCart) {
+    const orderToCreate = this.getOrderToCreate(shoppingCart!);
+    return firstValueFrom(this.checkoutService.createOrder(orderToCreate));
+  }
+
+  private async confirmPaymentWithStripe(shoppingCart: IShoppingCart) {
+    return this.stripe.confirmCardPayment(shoppingCart?.clientSecret, {
+      payment_method: {
+        card: this.cardNumber,
+        billing_details: {
+          name: this.checkoutForm.get('paymentForm')?.get('nameOnCard')?.value
+        }
+      }
+    });
+  }
 }
